chore: Rename "background" to "commands" where relevant, to avoid confusion with FastAPI background tasks. Organize code accordingly.
